#scandipwa-3616 - min max qty in wishlist support
